Refaktoriserat bort Anstalld.java
* Flyttat flera metoder från Anstalld till AnstalldMeny
* Uppdaterat AvdelningMeny och EditAvdelning så de pekar rätt. 
* Raderat Anstalld.java
